add go modules create packages some refactor
