+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/mike0sv/go-markov-bot
+
+go 1.14
+
+require (
+	github.com/influxdata/platform v0.0.0-20190117200541-d500d3cf5589
+	github.com/urfave/cli v1.22.4
+)
